added AdminController and moved all admin functionality there. updated garden and map controllers to use /user root. Changed templates to use new controller mapping. updated some owner and feature controllers to redirect from admin-list to their list files.
