+++ src/main/java/org/launchcode/water_garden_tour/controllers/HomeController.java
@@ -1,9 +1,8 @@
 package org.launchcode.water_garden_tour.controllers;
+
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 public class HomeController {
@@ -17,21 +16,5 @@ public class HomeController {
     }
 
 
-    @GetMapping("/home")
-    @ResponseBody
-    public String home() {
-        return "Welcome";
-    }
-
-    @GetMapping("/user")
-    @ResponseBody
-    public String user() {
-        return "Welcome User";
-    }
-
-    @GetMapping("/admin")
-    public String renderAdminPage() {
-        return "admin";
-    }
 
 }
