+++ src/main/resources/templates/features/add.html
@@ -9,7 +9,7 @@
 
 <div th:replace="fragments :: page-header"></div>
 
-<form method="post" th:action="@{/features/add}">
+<form method="post" th:action="@{/admin/features/add}">
     <div class="form-group">
         <label th:for="name">Feature Name</label>
         <input class="form-control" th:field="${feature.name}" />
