+++ src/main/resources/templates/features/update.html
@@ -1,15 +1,13 @@
 <!DOCTYPE html>
 <html lang="en" xmlns:th="http://www.thymeleaf.org/">
 <head th:replace="fragments :: head">
-    <meta charset="UTF-8"/>
-    <title></title>
 </head>
 
 <body>
 
 <div th:replace="fragments :: page-header"></div>
 
-<form method="POST" th:action="@{/features/update}" th:object="${feature}">
+<form method="POST" th:action="@{/admin/features/update}" th:object="${feature}">
     <input type="hidden" th:name="featureId" th:value="${feature.id}" />
     <div class="form-group">
         <label th:for="name">Feature Name</label>
