+++ src/main/resources/templates/gardens/update.html
@@ -9,7 +9,7 @@
 
 <div th:replace="fragments :: page-header"></div>
 
-<form method="POST" enctype="multipart/form-data" th:action="@{/gardens/update}" th:object="${garden}">
+<form method="POST" enctype="multipart/form-data" th:action="@{/admin/gardens/update}" th:object="${garden}">
     <input type="hidden" th:name="gardenId" th:value="${garden.id}" />
     <div class="form-group">
         <label th:for="name">Garden Name</label>
@@ -52,7 +52,7 @@
                 </option>
             </th:block>
         </select>
-        <a th:href="@{'/owners/add'}">Add Owner</a>
+        <a th:href="@{'/admin/owners/add'}">Add Owner</a>
     </div>
     <div class="form-group">
         <label th:for="features">Features: </label>
@@ -68,7 +68,7 @@
                    th:unless="${gardenFeatures.contains(feature)}">
             <label th:text="${feature.name}"></label>
         </th:block>
-        <a th:href="@{'/features/add'}">Add Features</a>
+        <a th:href="@{'/admin/features/add'}">Add Features</a>
         <div class="form-group">
             <label>Upload Image</label>
             <input class="form-control-file" type="file" name="file"/>
