+++ src/main/resources/templates/owners/update.html
@@ -9,7 +9,7 @@
 
 <div th:replace="fragments :: page-header"></div>
 
-<form method="POST" th:action="@{/owners/update}" th:object="${owner}">
+<form method="POST" th:action="@{/admin/owners/update}" th:object="${owner}">
     <input type="hidden" th:name="ownerId" th:value="${owner.id}" />
     <div class="form-group">
         <label th:for="name">Owner Name</label>
